src/statistics_service/requirements.txt
kafka-python==2.0.2
clickhouse-driver==0.2.8
Flask>=3.0.2
grpcio==1.40.0
protobuf==3.20